refactor: refatora caminhos e rotas
