+++ src/routes/routes.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from "react-router-dom";
-import { Timeline } from "./pages/Timeline/Timeline";
-import { Status } from "./pages/Status/Status";
-import { Default } from "./layouts/Default";
+import { Timeline } from "../pages/Timeline/Timeline";
+import { Status } from "../pages/Status/Status";
+import { Default } from "../layouts/Default";
 
 export const router = createBrowserRouter([
     {
